requirements.txt
pytest
selenium
webdriver_manager
pyyaml
jinja2
